src/backend/pyproject.toml
[project]
name = "helmos-agentic-layer"
version = "0.1.0"
description = "Production-oriented backend foundation for the HelmOS agentic layer."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aioboto3>=13.1.1",
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "greenlet>=3.1.1",
  "httpx>=0.28.1",
  "langchain>=0.3.20",
  "langgraph>=0.2.74",
  "langsmith>=0.3.15",
  "openai>=1.68.2",
  "opentelemetry-api>=1.32.1",
  "pgvector>=0.3.6",
  "psycopg[binary]>=3.2.6",
  "pydantic-settings>=2.7.1",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.37",
  "structlog>=25.1.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.3",
  "ruff>=0.9.7",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"
